requirements.txt
ffpyplayer
openai-whisper
ffmpeg-python
pytube
ffmpeg
nltk
google-generativeai
python-dotenv 